backend-flask/requirements.txt
flask
flask-cors

#Honeycomb
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests

#X-Ray
aws-xray-sdk

#CloudWatch
watchtower

#Rollbar
blinker
rollbar

Flask-AWSCognito

requests
python-jose

psycopg[binary,pool]

python-dotenv

boto3
python-dateutil

psycopg_pool